lab4/backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
flask-cors==3.0.10